+++ src/Facebook/FacebookBatchResponse.php
@@ -21,7 +21,7 @@
  * DEALINGS IN THE SOFTWARE.
  *
  */
-namespace Facebook\Entities;
+namespace Facebook;
 
 use ArrayIterator;
 use IteratorAggregate;