+++ tests/Helpers/FacebookRedirectLoginHelperTest.php
@@ -25,7 +25,7 @@ namespace Facebook\Tests\Helpers;
 
 use Mockery as m;
 use Facebook\Facebook;
-use Facebook\Entities\FacebookApp;
+use \Facebook\FacebookApp;
 use Facebook\Helpers\FacebookRedirectLoginHelper;
 use Facebook\PersistentData\FacebookMemoryPersistentDataHandler;
 use Facebook\PseudoRandomString\PseudoRandomStringGeneratorInterface;
@@ -101,7 +101,7 @@ class FacebookRedirectLoginHelperTest extends \PHPUnit_Framework_TestCase
     $_GET['state'] = 'foo_state';
     $_GET['code'] = 'foo_code';
 
-    $response = m::mock('Facebook\Entities\FacebookResponse');
+    $response = m::mock('Facebook\FacebookResponse');
     $response
       ->shouldReceive('getDecodedBody')
       ->once()
@@ -112,7 +112,7 @@ class FacebookRedirectLoginHelperTest extends \PHPUnit_Framework_TestCase
     $client = m::mock('Facebook\FacebookClient');
     $client
       ->shouldReceive('sendRequest')
-      ->with(m::type('Facebook\Entities\FacebookRequest'))
+      ->with(m::type('Facebook\FacebookRequest'))
       ->once()
       ->andReturn($response);
 
@@ -121,7 +121,7 @@ class FacebookRedirectLoginHelperTest extends \PHPUnit_Framework_TestCase
 
     $accessToken = $helper->getAccessToken($client, self::REDIRECT_URL);
 
-    $this->assertInstanceOf('Facebook\Entities\AccessToken', $accessToken);
+    $this->assertInstanceOf('Facebook\AccessToken', $accessToken);
     $this->assertEquals('access_token_from_code', (string) $accessToken);
   }